Refactor navigator facade layering and split manual scenarios
