+++ manual/__init__.py
@@ -1,0 +1,25 @@
+"""Manual scenarios and utilities for exploratory testing."""
+
+from .alarm import override, reliance
+from .gateway import commerce, fragments, translation, wording
+from .history import absence, surface
+from .navigator import siren
+from .tail import decline
+from .view import assent, rebuff, refuse, veto
+
+__all__ = [
+    "absence",
+    "assent",
+    "commerce",
+    "decline",
+    "fragments",
+    "rebuff",
+    "refuse",
+    "reliance",
+    "siren",
+    "surface",
+    "veto",
+    "wording",
+    "translation",
+    "override",
+]
